Struct Dependency Order Resolution
Struct are regexed out of the type names for the struct fields, with the
assumption that all dependent types will start with "vk", "VK", or "Vk".

Then, once all of the structs are extracted from the soup, they are "committed"
in order.  Valid dependencies will be recursively inserted before the struct
requiring them.  This ensures that the _fields_ are set in a legal ordering.