Iterating on Command Generation
This change adds some version defines, a broken system for lazy loading
instance and device commands, and a test sample code thing to validate
against.

vkGetInstanceProcAddr(instance, b'vkDestroySurfaceKHR') blows up with a
ffi_prep_cif error.  The instance is a VkInstance object constructed by
vkCreateInstance.

Seeing that sdl2.vulkan.VkInstance looks to effectively be a void pointer
type, I think my definition of VkInstance as an empty struct is wrong.

I'd love to be rid of this weird indirection for instance and device
functions.