Enums Support Arithmetic Operators
This makes it possible to do stuff like this:

    VK_QUEUE_GRAPHICS_BIT | VK_QUEUE_COMPUTE_BIT | VK_QUEUE_TRANSFER_BIT

I'm considering ripping this out in favor of making enums values all by python
integers.  The main reason is assigning values to VkFlag (ctypes.c_uint32)
struct members works fine for python numbers, but blows up for other types.

So, that leads to code that looks like this and confusing error messages:

    aspectMask = VK_IMAGE_ASPECT_COLOR_BIT.value

The enum classes themselves (eg VkFormat) should be translated to ctypes.c_int
when used in structs etc, though it might be useful to define these as
separate types.  It would be nice to use the enum class name for python enum
classes containing the named enumerations, instead of the "name" dict hack.

The other change I'm thinking of making is VkFlags inheriting from an
intermediary to c_uint32, so as to provide bitwise operators etc so you don't
have to do stuff like this:

    if (queue_family.queueFlags.value & match)

I'm not thrilled by the idea of subclassing every numeric ctype I want to
support common arithmetic operatons on, but there's not really much of a
choice for types used in struct fields.